+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Slug validation when adding a new affiliate
+- Form label name changes
+
 ## [1.3.4] - 2022-09-30
 
 ## [1.3.3] - 2022-09-30